chore: separated provider tests into different files. fixed some of its tests
